2nd draft of whitebox finished
I split the upper deck and the lower deck into 2 scenes since the ride in the dumbwaiter can surve as a scene loading point.
I also removed the kitchen since it was not needed in the storyline. The dunbwaiter is now located in the hall by boiler room and the top of it goes to a bar on the deck.