Fixed the lockbox to be more specific
The lock box  only checked for a specific amount of collectables instead of specific ones, this had the potential to cause false positives so I fixed it. The clue panel got deleted from scene so I remade it AND made it a prefab so it can be used in other scenes to show the clues for the player.